pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.79.0",
    "google-genai>=1.62.0",
    "openai>=2.17.0",
    "tenacity>=9.1.4",
]
